test(agents): split memory context tests into focused modules
