+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-21  
-> **293** tracked files
+> **294** tracked files
 
 ## Summary
 
@@ -9,7 +9,7 @@
 |---------------------|-------|
 | `docs/` | 67 |
 | `internal/` | 57 |
-| `agents/` | 48 |
+| `agents/` | 49 |
 | `tests/` | 36 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 96 |
+| `.py` | 97 |
 | `.md` | 83 |
 | `.go` | 53 |
 | `.yaml` | 23 |
@@ -127,8 +127,9 @@
 │   ├── task_agent.py
 │   ├── tests/
 │   │   ├── __init__.py
+│   │   ├── test_inject_memory_context.py
 │   │   ├── test_memory_budget.py
-│   │   └── test_persona_runtime_memory_context.py
+│   │   └── test_truncate_with_ellipsis.py
 │   ├── tick.py
 │   ├── tools/
 │   │   ├── __init__.py
